+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYPP1H6E3DSJGNG5T58YX4J5"
+        "01KYQ73NJZ4ZG3S6XPNTD1NKD4"
     )
     r"""The unique identifier of the data source"""