+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-29T10:16:55.110Z"
+    created: Optional[str] = "2026-07-29T15:15:11.668Z"
 
-    updated: Optional[str] = "2026-07-29T10:16:55.110Z"
+    updated: Optional[str] = "2026-07-29T15:15:11.668Z"
 
     updated_by_id: OptionalNullable[str] = UNSET