+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -59,9 +59,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-07-29T10:16:55.110Z"
+    created: Optional[str] = "2026-07-29T15:15:11.668Z"
 
-    updated: Optional[str] = "2026-07-29T10:16:55.110Z"
+    updated: Optional[str] = "2026-07-29T15:15:11.668Z"
 
     updated_by_id: OptionalNullable[str] = UNSET