+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYPP1H6BZHZP0GFV5GYSRFDT"
+        "01KYQ73NJXD8TREC6C0X43SCA1"
     )
     r"""The unique identifier of the data source"""