+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYPP1H6DSR6D22QX6RZXQKEP"
+        "01KYQ73NJYBCGTTJVTV9683KMW"
     )
     r"""The unique identifier of the data source"""