+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYPP1H6HN2QT9QB97H7TK9PX"
+        "01KYQ73NK24AXJ1CDFFAJ210HC"
     )
     r"""The unique identifier of the data source"""