+++ packages/orq-rc/RELEASES.md
@@ -9819,3 +9819,13 @@ Based on:
 - [python v4.13.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.14 - packages/orq-rc
+
+## 2026-07-29 15:18:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.16 - packages/orq-rc